requirements.txt
# ═══════════════════════════════════════════════════════════
# Telegram Video Downloader Bot - Dependencies
# ═══════════════════════════════════════════════════════════

# ═══════════════════════════════════════════════════════════
# Core Bot Dependencies - مكتبات البوت الأساسية
# ═══════════════════════════════════════════════════════════
pyrogram>=2.0.0
tgcrypto>=1.2.5
python-dotenv>=1.0.0

# ═══════════════════════════════════════════════════════════
# Database - قاعدة البيانات
# ═══════════════════════════════════════════════════════════
psycopg2-binary>=2.9.9

# ═══════════════════════════════════════════════════════════
# Video/Media Downloaders - أدوات تحميل الفيديو والوسائط
# ═══════════════════════════════════════════════════════════
yt-dlp>=2023.12.30
gallery-dl>=1.26.0
instaloader>=4.10.0

# ═══════════════════════════════════════════════════════════
# HTTP Requests - طلبات HTTP
# ═══════════════════════════════════════════════════════════
requests>=2.31.0
